fixed and updated Unittests

+++ Models/Chat.cs
@@ -3,8 +3,8 @@
     public class Chat
     {
         public int id { get; set; }
-        public int sender_id { get; set; }
-        public int receiver_id { get; set; }
+        public long sender_id { get; set; }
+        public long receiver_id { get; set; }
         public string topic { get; set; }
     }
 }
